refactor : added infrastructure directory for repository implementation
