+++ src/main/java/ulaval/glo2003/domain/infrastructure/inMemory/InMemoryOfferRepository.java
@@ -1,6 +1,9 @@
-package ulaval.glo2003.domain.offer;
+package ulaval.glo2003.domain.infrastructure.inMemory;
 
 import jakarta.ws.rs.NotFoundException;
+import ulaval.glo2003.domain.offer.IOfferRepository;
+import ulaval.glo2003.domain.offer.Offer;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
