+++ src/main/java/ulaval/glo2003/domain/infrastructure/inMemory/InMemoryProductRepository.java
@@ -1,6 +1,10 @@
-package ulaval.glo2003.domain.product;
+package ulaval.glo2003.domain.infrastructure.inMemory;
 
 import jakarta.ws.rs.NotFoundException;
+import ulaval.glo2003.domain.product.IProductRepository;
+import ulaval.glo2003.domain.product.Product;
+import ulaval.glo2003.domain.product.ProductFilter;
+
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
