+++ src/main/java/ulaval/glo2003/domain/infrastructure/mongo/MongoSellerRepository.java
@@ -1,9 +1,12 @@
-package ulaval.glo2003.domain.seller;
+package ulaval.glo2003.domain.infrastructure.mongo;
 
 import dev.morphia.Datastore;
 import dev.morphia.query.Query;
 import dev.morphia.query.filters.Filters;
 import jakarta.ws.rs.NotFoundException;
+import ulaval.glo2003.domain.seller.ISellerRepository;
+import ulaval.glo2003.domain.seller.Seller;
+
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
