+++ src/main/java/ulaval/glo2003/service/SellingServiceFactory.java
@@ -2,16 +2,16 @@ package ulaval.glo2003.service;
 
 import dev.morphia.Datastore;
 import ulaval.glo2003.domain.offer.IOfferRepository;
-import ulaval.glo2003.domain.offer.InMemoryOfferRepository;
-import ulaval.glo2003.domain.offer.MongoOfferRepository;
+import ulaval.glo2003.domain.infrastructure.inMemory.InMemoryOfferRepository;
+import ulaval.glo2003.domain.infrastructure.mongo.MongoOfferRepository;
 import ulaval.glo2003.domain.offer.OfferFactory;
 import ulaval.glo2003.domain.product.IProductRepository;
-import ulaval.glo2003.domain.product.InMemoryProductRepository;
-import ulaval.glo2003.domain.product.MongoProductRepository;
+import ulaval.glo2003.domain.infrastructure.inMemory.InMemoryProductRepository;
+import ulaval.glo2003.domain.infrastructure.mongo.MongoProductRepository;
 import ulaval.glo2003.domain.product.ProductFactory;
 import ulaval.glo2003.domain.seller.ISellerRepository;
-import ulaval.glo2003.domain.seller.InMemorySellerRepository;
-import ulaval.glo2003.domain.seller.MongoSellerRepository;
+import ulaval.glo2003.domain.infrastructure.inMemory.InMemorySellerRepository;
+import ulaval.glo2003.domain.infrastructure.mongo.MongoSellerRepository;
 import ulaval.glo2003.domain.seller.SellerFactory;
 
 public class SellingServiceFactory {
