+++ src/test/java/ulaval/glo2003/domain/infrastructure/inMemory/InMemoryOfferRepositoryTest.java
@@ -1,4 +1,7 @@
-package ulaval.glo2003.domain.offer;
+package ulaval.glo2003.domain.infrastructure.inMemory;
+
+import ulaval.glo2003.domain.offer.IOfferRepository;
+import ulaval.glo2003.domain.offer.IOfferRepositoryTest;
 
 class InMemoryOfferRepositoryTest extends IOfferRepositoryTest {
