+++ src/test/java/ulaval/glo2003/domain/infrastructure/inMemory/InMemoryProductRepositoryTest.java
@@ -1,4 +1,7 @@
-package ulaval.glo2003.domain.product;
+package ulaval.glo2003.domain.infrastructure.inMemory;
+
+import ulaval.glo2003.domain.product.IProductRepository;
+import ulaval.glo2003.domain.product.IProductRepositoryTest;
 
 class InMemoryProductRepositoryTest extends IProductRepositoryTest {
     @Override
