+++ src/test/java/ulaval/glo2003/domain/infrastructure/inMemory/InMemorySellerRepositoryTest.java
@@ -1,4 +1,7 @@
-package ulaval.glo2003.domain.seller;
+package ulaval.glo2003.domain.infrastructure.inMemory;
+
+import ulaval.glo2003.domain.seller.ISellerRepository;
+import ulaval.glo2003.domain.seller.ISellerRepositoryTest;
 
 public class InMemorySellerRepositoryTest extends ISellerRepositoryTest {
