+++ src/test/java/ulaval/glo2003/domain/infrastructure/mongo/MongoProductRepositoryTest.java
@@ -1,18 +1,21 @@
-package ulaval.glo2003.domain.offer;
+package ulaval.glo2003.domain.infrastructure.mongo;
 
 import com.mongodb.MongoClientSettings;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoClients;
 import dev.morphia.Datastore;
 import dev.morphia.Morphia;
+import ulaval.glo2003.domain.product.IProductRepository;
+import ulaval.glo2003.domain.product.IProductRepositoryTest;
+
 import java.util.concurrent.TimeUnit;
 
-public class MongoOfferRepositoryTest extends IOfferRepositoryTest {
+public class MongoProductRepositoryTest extends IProductRepositoryTest {
 
     private final int TIMEOUT = 5000;
 
     @Override
-    protected IOfferRepository createRepository() {
+    protected IProductRepository createRepository() {
         MongoClient client = MongoClients.create(MongoClientSettings.builder()
                 .applyToClusterSettings(builder -> builder.serverSelectionTimeout(TIMEOUT, TimeUnit.MILLISECONDS))
                 .applyToConnectionPoolSettings(builder -> builder.maxConnectionIdleTime(TIMEOUT, TimeUnit.MILLISECONDS))
@@ -21,6 +24,6 @@ public class MongoOfferRepositoryTest extends IOfferRepositoryTest {
         datastore.getMapper().mapPackage("ulaval.glo2003");
         datastore.ensureIndexes();
 
-        return new MongoOfferRepository(datastore);
+        return new MongoProductRepository(datastore);
     }
 }
